NETOBSERV-1517: remove IPFIX agent mode
- Remove controller code for OVS configmap / CNO configuration
- Remove related tests (and move console plugin related tests to its
  dedicated file)
- Remove FLP ingester reconciler (was only used with ipfix)
- Note that the IPFIX api is not removed to not introduce breaking
  changes in existing API, however it is now documented that using IPFIX
would have no effect
