+++ apis/flowcollector/v1beta2/flowcollector_types.go
@@ -93,12 +93,10 @@ const (
 // allow defining both fields.
 // +union
 type FlowCollectorAgent struct {
-	// `type` selects the flows tracing agent. Possible values are:<br>
-	// - `eBPF` (default) to use NetObserv eBPF agent.<br>
-	// - `IPFIX` [deprecated (*)] - to use the legacy IPFIX collector.<br>
-	// `eBPF` is recommended as it offers better performances and should work regardless of the CNI installed on the cluster.
-	// `IPFIX` works with OVN-Kubernetes CNI (other CNIs could work if they support exporting IPFIX,
-	// but they would require manual configuration).
+	// `type` [deprecated (*)] selects the flows tracing agent. The only possible value is `eBPF` (default), to use NetObserv eBPF agent.<br>
+	// Previously, using an IPFIX collector was allowed, but was deprecated and it is now removed.<br>
+	// Setting `IPFIX` is ignored and still use the eBPF Agent.
+	// Since there is only a single option here, this field will be remove in a future API version.
 	// +unionDiscriminator
 	// +kubebuilder:validation:Enum:="eBPF";"IPFIX"
 	// +kubebuilder:default:=eBPF
@@ -920,7 +918,7 @@ type AdvancedProcessorConfig struct {
 
 	//+kubebuilder:default:=true
 	//+optional
-	// `dropUnusedFields` allows, when set to `true`, to drop fields that are known to be unused by OVS, to save storage space.
+	// `dropUnusedFields` [deprecated (*)] this setting is not used anymore.
 	DropUnusedFields *bool `json:"dropUnusedFields,omitempty"`
 
 	//+kubebuilder:default:="30s"
