+++ controllers/suite_test.go
@@ -50,7 +50,6 @@ func TestAPIs(t *testing.T) {
 // go test ./... runs always Ginkgo test suites in parallel and they would interfere
 // this way we make sure that both test sub-suites are executed serially
 var _ = Describe("FlowCollector Controller", Ordered, Serial, func() {
-	flowCollectorControllerSpecs()
 	flowCollectorConsolePluginSpecs()
 	flowCollectorEBPFSpecs()
 	flowCollectorEBPFKafkaSpecs()
